Organized some tests
