+++ pkg/request/json_test.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/hellofresh/janus/pkg/mock"
 	"github.com/hellofresh/janus/pkg/request"
+	"github.com/hellofresh/janus/pkg/test"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -15,19 +15,19 @@ func TestBindSimpleJSON(t *testing.T) {
 	req, _ := http.NewRequest("GET", "/", nil)
 	req.Body = ioutil.NopCloser(bytes.NewBuffer([]byte("{\"name\": \"Test Recipe\", \"tags\": [\"test\"]}")))
 
-	recipe := mock.Recipe{}
+	recipe := test.Recipe{}
 	err := request.BindJSON(req, &recipe)
 
 	assert.Nil(t, err)
 	assert.Equal(t, "Test Recipe", recipe.Name)
-	assert.Equal(t, []mock.Tag{"test"}, recipe.Tags)
+	assert.Equal(t, []test.Tag{"test"}, recipe.Tags)
 }
 
 func TestWrongBindJSON(t *testing.T) {
 	req, _ := http.NewRequest("GET", "/", nil)
 	req.Body = ioutil.NopCloser(bytes.NewBuffer([]byte{}))
 
-	recipe := mock.Recipe{}
+	recipe := test.Recipe{}
 	err := request.BindJSON(req, &recipe)
 
 	assert.NotNil(t, err)
